+ Bomb of code to rename clientException to clientError + Use authenticated user in ClientErrorService
